--- mail_debrand/models/mail_thread.py
@@ -1,14 +1,0 @@
-# Copyright 2019 Eficent Business and IT Consulting Services S.L.
-#   Lois Rilo <lois.rilo@eficent.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from odoo import models
-
-
-class MailThread(models.AbstractModel):
-    _inherit = "mail.thread"
-
-    def _replace_local_links(self, html, base_url=None):
-        html = super()._replace_local_links(html, base_url=base_url)
-        html_debranded = self.env["mail.template"]._debrand_body(html)
-        return html_debranded
